requirements.txt
# Core
python-dotenv
psycopg  # ← Para PostgresSaver (versión 3)
psycopg2-binary  # ← Para tu config/database.py
python-telegram-bot

# LangChain y LangGraph
langchain
langchain-core
langchain-anthropic  # o langchain-openai según tu LLM
langgraph
langgraph-checkpoint-postgres  # ← IMPORTANTE: para memoria persistente

# Tools adicionales
langchain-community  # si usas TavilySearch u otras tools